requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
mlflow>=2.10.0
faker>=24.0.0
loguru>=0.7.0
pyyaml>=6.0
streamlit>=1.32.0
plotly>=5.18.0
openpyxl>=3.1.0
pytest>=7.4.0
httpx>=0.26.0
ruff>=0.3.0
